pyproject.toml
[tool.poetry]
name = "file_cache"
version = "0.1.1"
description = ""
authors = []
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.63.0"
python-dotenv = "^0.15.0"
aiofiles = "^0.6.0"
python-multipart = "^0.0.5"
uvloop = { version = "^0.14.0", markers = "sys_platform != 'win32' and implementation_name == 'cpython'", optional = true }
aioredis = { version = "^1.3.1", optional = true }


[tool.poetry.dev-dependencies]
flake8 = "*"
isort = "*"
black = "^20.8b1"
pytest = "*"
pytest-xdist = "*"
pytest-cov = "*"
pytest-asyncio = "*"
asynctest = "*"
bandit = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# 谨慎使用，会导致包更新不及时
[[tool.poetry.source]]
url = "https://mirrors.aliyun.com/pypi/simple/"
name = "aliyun"
default = true

[tool.poetry.extras]
aioredis = ["aioredis"]
uvloop = ["uvloop"]